4th b.c

+++ ABC/200/b.c
@@ -2,8 +2,8 @@
 
 int main()
 {
-      int N;
-      int K;
+    unsigned long N;
+    unsigned long K;
     scanf("%d", &N);
     scanf("%d", &K);
     for (int i = 0; i < K; i++)
